Fixed issue with empty record response from DNS API
The API was not returning any records when using the subdomain version of a tld. (try.fluxvpn) which what was tied to `${domain_name}`

The API attempts to find subdomains of the input record: `*.try.fluxvpn` which didn't actually exist.

Changing the domain name parameter to the zone name parameter in the API call on line 126 allows for the endpoint to pull all records for all domains in the created zone. `${zone_name}`